Extract LRO status retrieval into a helper function.
